+++ kubernetes-server-mock/src/main/java/io/fabric8/kubernetes/client/server/mock/KubernetesCrudDispatcher.java
@@ -269,10 +269,14 @@ public class KubernetesCrudDispatcher extends CrudDispatcher {
       query = query.add(new Attribute("name", resourceName));
     }
     WatchEventsListener watchEventListener = new WatchEventsListener(context, query, watchEventListeners, LOGGER,
-      watch -> 
-        map.entrySet().stream()
-          .filter(entry -> watch.attributeMatches(entry.getKey()))
-          .forEach(entry -> watch.sendWebSocketResponse(entry.getValue(), Action.ADDED)));
+      watch -> {
+        synchronized (KubernetesCrudDispatcher.this) {
+          map.entrySet().stream()
+            .filter(entry -> watch.attributeMatches(entry.getKey()))
+            .forEach(entry -> watch.sendWebSocketResponse(entry.getValue(), Action.ADDED));
+        }
+      }
+    );
     watchEventListeners.add(watchEventListener);
     mockResponse.setSocketPolicy(SocketPolicy.KEEP_OPEN);
     return mockResponse.withWebSocketUpgrade(watchEventListener);